backend/requirements.txt
asyncio
numpy<2.0.0
tiktoken==0.7.0
python-dotenv==1.0.1
openai==1.47.1
scikit-learn==1.5.2
tenacity>=8.2.0,<8.4.0
black==24.8.0
llama_index
llama_hub[reddit,google_gsheets,twitter]
llama-index-core==0.11.12
llama-index-readers-google==0.4.1
llama-index-readers-web==0.2.2
llama-index-readers-notion==0.2.0
llama-index-readers-reddit==0.2.0
llama-index-readers-twitter==0.2.0
llama-index-readers-wikipedia==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-json==0.2.0
elevenlabs==1.9.0
fal-client==0.4.1
llama-index-readers-github==0.2.0
arxiv==2.1.3
pymupdf==1.24.10
fastapi
uvicorn
pydantic
apscheduler
sqlalchemy
dask
distributed
bokeh>=3.1.0
attrs
Pillow
scipy
nest_asyncio
